+++ src/atom/index.ts
@@ -1,0 +1,6 @@
+import Atom from "./atom";
+import Atomizer, { AtomizerTemplate } from "./atomizer";
+
+export {
+    Atom, Atomizer, AtomizerTemplate
+}