+++ src/components/index.ts
@@ -1,0 +1,11 @@
+import Component from "./component";
+import Script from "./scripts";
+import String from "./string";
+import Style from "./styles";
+
+export {
+    Component,
+    Script,
+    String,
+    Style
+};