+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+# [1.0.1-alpha.3] - 30-03-2023
+
+### Added
+- Atom Class.
+- Atomizer Class.
+- AtomizerTemplate Class.
+- templates as `[name].component.html` files in `src/templates/` folder.
+
+### Changed
+- The entire `solaris.ts` got overhauled.
+- `Logger` class saves logs under `./logs` folder.
+- `basic` folder is renamed to `components`
+- `logger`,`filemanager`, `solaris` moved into `./src/utils` folder.
+### Removed 
+- Almost all `.ts` files that Inherited from `Component` class.
+- Init method from `SolarisUI` class.
 
 # [1.0.1-alpha.2] - 25-03-2023
 ### Added